refactor: extract shared helpers for availability, EKS workloads, SQL flow, and AWS credentials
