+++ app/tools/EKSEventsTool/__init__.py
@@ -6,8 +6,8 @@ import logging
 from typing import Any, cast
 
 from app.services.eks.eks_k8s_client import build_k8s_clients
-from app.tools.EKSListClustersTool import _eks_creds
 from app.tools.tool_decorator import tool
+from app.tools.utils.aws import aws_creds
 from app.tools.utils.availability import eks_available_or_backend
 
 logger = logging.getLogger(__name__)
@@ -23,7 +23,7 @@ def _events_extract_params(sources: dict[str, dict]) -> dict[str, Any]:
         "cluster_name": eks.get("cluster_name", ""),
         "namespace": eks.get("namespace", "default"),
         "eks_backend": eks.get("_backend"),
-        **_eks_creds(eks),
+        **aws_creds(eks),
     }
