+++ app/tools/MySQLCurrentProcessesTool/__init__.py
@@ -9,6 +9,7 @@ from app.integrations.mysql import (
     resolve_mysql_config,
 )
 from app.tools.tool_decorator import tool
+from app.tools.utils.sql import sql_tool_flow
 
 
 @tool(
@@ -31,13 +32,11 @@ def get_mysql_current_processes(
     port: int = 3306,
 ) -> dict[str, Any]:
     """Fetch active processes running longer than threshold_seconds (default 1s)."""
-    _db_defaulted = database is None
-    if database is None:
-        database = "mysql"
-    config = resolve_mysql_config(host=host, database=database, port=port)
-    result = get_current_processes(config, threshold_seconds=threshold_seconds)
-    if _db_defaulted:
-        result["default_db_warning"] = (
-            "WARNING: No database was specified; defaulted to 'mysql'. Results may not reflect application data."
-        )
-    return result
+    return sql_tool_flow(
+        database=database,
+        default_db="mysql",
+        resolve_func=resolve_mysql_config,
+        execute_func=get_current_processes,
+        identifying_params={"host": host, "port": port},
+        execute_params={"threshold_seconds": threshold_seconds},
+    )
